211 : Add missing unit tests
